move CredentialResponseCoordinator into utils
