+++ packages/core/src/modules/credentials/index.ts
@@ -7,4 +7,3 @@ export * from './CredentialState'
 export * from './CredentialEvents'
 export * from './CredentialAutoAcceptType'
 export * from './CredentialProtocolVersion'
-export * from './CredentialResponseCoordinator'
