+++ packages/core/src/modules/credentials/protocol/v1/V1CredentialService.ts
@@ -33,6 +33,7 @@ import { EventEmitter } from '../../../../agent/EventEmitter'
 import { ServiceDecorator } from '../../../../decorators/service/ServiceDecorator'
 import { AriesFrameworkError } from '../../../../error'
 import { DidCommMessageRepository, DidCommMessageRole } from '../../../../storage'
+import { CredentialResponseCoordinator } from '../../../../utils/CredentialResponseCoordinator'
 import { isLinkedAttachment } from '../../../../utils/attachment'
 import { AckStatus } from '../../../common'
 import { ConnectionService } from '../../../connections/services'
@@ -40,7 +41,6 @@ import { MediationRecipientService } from '../../../routing'
 import { AutoAcceptCredential } from '../../CredentialAutoAcceptType'
 import { CredentialEventTypes } from '../../CredentialEvents'
 import { CredentialProtocolVersion } from '../../CredentialProtocolVersion'
-import { CredentialResponseCoordinator } from '../../CredentialResponseCoordinator'
 import { CredentialState } from '../../CredentialState'
 import { CredentialUtils } from '../../CredentialUtils'
 import { CredentialProblemReportError, CredentialProblemReportReason } from '../../errors'
